# doc/apidoc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/matches": {
      "get": {
        "operationId": "MatchesController_list",
        "parameters": [
          {
            "name": "tournament",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "org",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}": {
      "get": {
        "operationId": "MatchesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Match detail + innings summary.",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/state": {
      "get": {
        "operationId": "MatchesController_state",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Live state snapshot — Redis-first, Postgres fallback (`source` field says which). Prefer the WebSocket feed (/live) for push updates.",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/balls/recent": {
      "get": {
        "operationId": "MatchesController_recentBalls",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "Last N ball events (Redis stream) for instant UI hydration.",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/presence": {
      "get": {
        "operationId": "MatchesController_presence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Live presence: {viewers, scorers} currently connected via WebSocket.",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/scorecard": {
      "get": {
        "operationId": "MatchesController_scorecard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "Full scorecard (batting/bowling cards, fall of wickets).",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/overs": {
      "get": {
        "operationId": "MatchesController_overs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "Over-by-over data (Manhattan / over comparison).",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/commentary": {
      "get": {
        "operationId": "MatchesController_commentary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "before",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches & Scoring"
        ]
      },
      "post": {
        "operationId": "MatchesController_addCommentary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentaryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Add manual commentary (commentator/scorer).",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/mvp": {
      "get": {
        "operationId": "MatchesController_mvp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/stats": {
      "get": {
        "operationId": "MatchesController_matchStats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Stats tab: wagon-wheel vectors, partnerships, over-by-over run rate.",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/squads": {
      "get": {
        "operationId": "MatchesController_squads",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches & Scoring"
        ]
      },
      "put": {
        "operationId": "MatchesController_setSquad",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSquadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "Set one team's squad: playing XI + 12th man with can_bat/can_bowl toggles.",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/teams/{teamA}/head-to-head/{teamB}": {
      "get": {
        "operationId": "MatchesController_headToHead",
        "parameters": [
          {
            "name": "teamA",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teamB",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/tournaments/{id}/stats/leaders": {
      "get": {
        "operationId": "MatchesController_leaders",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "Tournament leaderboards: metric = runs | wickets | mvp | sr | economy",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/orgs/{orgId}/matches": {
      "post": {
        "operationId": "MatchesController_create",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/scorers": {
      "post": {
        "operationId": "MatchesController_assignScorer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignScorerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Grant a user match-scoped scorer access.",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/officials": {
      "put": {
        "operationId": "MatchesController_setOfficials",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetOfficialsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Set match officials (umpires, TV umpire, referee, scorer).",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/substitutions": {
      "post": {
        "operationId": "MatchesController_substitute",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubstitutionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Substitution: 12th man / concussion / impact player, with bat-bowl eligibility.",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/toss": {
      "post": {
        "operationId": "MatchesController_toss",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TossDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
    "/api/v1/matches/{id}/openers": {
      "post": {
        "operationId": "MatchesController_openers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Set opening batters + bowler; match goes live. Also used after an innings break.",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
...
    "/api/v1/matches/{id}/balls/last": {
      "delete": {
        "operationId": "MatchesController_undo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Undo the most recent ball (append-only supersede + full replay rebuild).",
        "tags": [
          "Matches & Scoring"
        ]
      }
    },
...
  },
  "info": {
    "title": "CricLive API",
    "description": "Cricket live scoring platform ? auth, organizations, tournaments, ball-by-ball scoring, stats, SaaS plans, CMS. Authorize with the access_token from POST /auth/login.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste access_token from /auth/login"
      }
    },
    "schemas": {
...
      "CreateMatchDto": {
        "type": "object",
        "properties": {}
      },
      "SetSquadDto": {
        "type": "object",
        "properties": {}
      },
      "AssignScorerDto": {
        "type": "object",
        "properties": {}
      },
      "SetOfficialsDto": {
        "type": "object",
        "properties": {}
      },
      "SubstitutionDto": {
        "type": "object",
        "properties": {}
      },
      "TossDto": {
        "type": "object",
        "properties": {}
      },
      "OpenersDto": {
        "type": "object",
        "properties": {}
      },
...
      "CommentaryDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}